+++ skops/io/_trusted_types.py
@@ -2,7 +2,7 @@ import numpy as np
 import scipy
 from sklearn.utils import all_estimators
 
-from ._utils import get_type_name
+from ._utils import get_public_type_names, get_type_name
 
 PRIMITIVES_TYPES = [int, float, str, bool]
 
@@ -14,35 +14,15 @@ SKLEARN_ESTIMATOR_TYPE_NAMES = [
     if get_type_name(estimator_class).startswith("sklearn.")
 ]
 
-SCIPY_UFUNC_TYPE_NAMES = sorted(
-    set(
-        [
-            get_type_name(getattr(scipy.special, attr))
-            for attr in dir(scipy.special)
-            if isinstance(getattr(scipy.special, attr), np.ufunc)
-            and get_type_name(getattr(scipy.special, attr)).startswith("scipy")
-        ]
-    )
-)
+SCIPY_UFUNC_TYPE_NAMES = get_public_type_names(module=scipy.special, _type=np.ufunc)
 
-NUMPY_UFUNC_TYPE_NAMES = sorted(
-    set(
-        [
-            get_type_name(getattr(np, attr))
-            for attr in dir(np)
-            if isinstance(getattr(np, attr), np.ufunc)
-            and get_type_name(getattr(np, attr)).startswith("numpy")
-        ]
-    )
-)
+NUMPY_UFUNC_TYPE_NAMES = get_public_type_names(module=np, _type=np.ufunc)
 
 NUMPY_DTYPE_TYPE_NAMES = sorted(
-    set(
-        [
-            get_type_name(dtype)
-            for dtypes in np.sctypes.values()
-            for dtype in dtypes  # type: ignore
-            if get_type_name(dtype).startswith("numpy")
-        ]
-    )
+    {
+        type_name
+        for dtypes in np.sctypes.values()
+        for dtype in dtypes  # type: ignore
+        if (type_name := get_type_name(dtype)).startswith("numpy")
+    }
 )